Move UnitTester to its Chisel3 home. Move GCDUnitTest.scala up so we can convert to run under TestSuite with the rest of the Scala/JUnit tests.
